Split Chat and ChatMessageList
